Refactor NavigationDrawer to be generic
