+++ client/src/components/List/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./List";
+export * from "./List";
